Fix tenant confusion -- workers arent tenant specific and shouldn't need to track task tenancy
