+++ tests/grpc_expedite_tests.rs
@@ -94,7 +94,6 @@ async fn grpc_expedite_future_scheduled_job() -> anyhow::Result<()> {
             .report_outcome(ReportOutcomeRequest {
                 shard: 0,
                 task_id: task.id.clone(),
-                tenant: None,
                 outcome: Some(report_outcome_request::Outcome::Success(MsgpackBytes {
                     data: rmp_serde::to_vec(&serde_json::json!("expedited success")).unwrap(),
                 })),
@@ -232,7 +231,6 @@ async fn grpc_expedite_running_job_fails() -> anyhow::Result<()> {
             .report_outcome(ReportOutcomeRequest {
                 shard: 0,
                 task_id: task.id.clone(),
-                tenant: None,
                 outcome: Some(report_outcome_request::Outcome::Success(MsgpackBytes {
                     data: rmp_serde::to_vec(&serde_json::json!("done")).unwrap(),
                 })),
