+++ tests/turmoil_runner/scenarios/concurrency_limits.rs
@@ -392,7 +392,6 @@ pub fn run() {
                             match client
                                 .report_outcome(tonic::Request::new(ReportOutcomeRequest {
                                     shard: 0,
-                                    tenant: None,
                                     task_id: task.id.clone(),
                                     outcome: Some(outcome),
                                 }))
@@ -461,7 +460,6 @@ pub fn run() {
                     match client
                         .report_outcome(tonic::Request::new(ReportOutcomeRequest {
                             shard: 0,
-                            tenant: None,
                             task_id: task.id.clone(),
                             outcome: Some(report_outcome_request::Outcome::Success(MsgpackBytes {
                                 data: rmp_serde::to_vec(&serde_json::json!("done")).unwrap(),
