+++ tests/turmoil_runner/scenarios/expedite_concurrency.rs
@@ -341,7 +341,6 @@ pub fn run() {
                     match client
                         .report_outcome(tonic::Request::new(ReportOutcomeRequest {
                             shard: 0,
-                            tenant: None,
                             task_id: task.id.clone(),
                             outcome: Some(report_outcome_request::Outcome::Success(MsgpackBytes {
                                 data: rmp_serde::to_vec(&serde_json::json!("done")).unwrap(),
