+++ tests/turmoil_runner/scenarios/grpc_end_to_end.rs
@@ -107,7 +107,6 @@ pub fn run() {
                 client
                     .report_outcome(tonic::Request::new(ReportOutcomeRequest {
                         shard: 0,
-                        tenant: None,
                         task_id: task.id.clone(),
                         outcome: Some(report_outcome_request::Outcome::Success(MsgpackBytes {
                             data: rmp_serde::to_vec(&serde_json::json!({"result": "ok"})).unwrap(),
