+++ tests/turmoil_runner/scenarios/high_message_loss.rs
@@ -133,7 +133,6 @@ pub fn run() {
                             match client
                                 .report_outcome(tonic::Request::new(ReportOutcomeRequest {
                                     shard: 0,
-                                    tenant: None,
                                     task_id: task.id.clone(),
                                     outcome: Some(report_outcome_request::Outcome::Success(
                                         MsgpackBytes {
