+++ tests/turmoil_runner/scenarios/retry_releases_ticket.rs
@@ -193,7 +193,6 @@ pub fn run() {
                             client
                                 .report_outcome(tonic::Request::new(ReportOutcomeRequest {
                                     shard: 0,
-                                    tenant: None,
                                     task_id: task.id.clone(),
                                     outcome: Some(report_outcome_request::Outcome::Failure(
                                         silo::pb::Failure {
@@ -233,7 +232,6 @@ pub fn run() {
                             client
                                 .report_outcome(tonic::Request::new(ReportOutcomeRequest {
                                     shard: 0,
-                                    tenant: None,
                                     task_id: task.id.clone(),
                                     outcome: Some(report_outcome_request::Outcome::Success(
                                         MsgpackBytes {
@@ -313,7 +311,6 @@ pub fn run() {
                         client
                             .report_outcome(tonic::Request::new(ReportOutcomeRequest {
                                 shard: 0,
-                                tenant: None,
                                 task_id: task.id.clone(),
                                 outcome: Some(report_outcome_request::Outcome::Success(
                                     MsgpackBytes {
